add some service

+++ src/main/java/com/dinner/Whatistomorrowfordinner/model/NutritionPlanData.java
@@ -2,16 +2,16 @@ package com.dinner.Whatistomorrowfordinner.model;
 
 public record NutritionPlanData(
         String planName,
-        int kcal,
-        int numberOfDays,
+        long kcal,
+        long numberOfDays,
         boolean breakfast,
         boolean lunch,
         boolean dinner,
         boolean snack,
         boolean supper,
-        int meal1,
-        int meal2,
-        int meal3,
-        int meal4,
-        int meal5) {
+        long meal1,
+        long meal2,
+        long meal3,
+        long meal4,
+        long meal5) {
 }
